refactor(PasswordInput): use stateless pw input
