+++ src/components/TextInput/TextInput.js
@@ -10,8 +10,7 @@ import React from 'react';
 import classNames from 'classnames';
 import { settings } from 'carbon-components';
 import WarningFilled16 from '@carbon/icons-react/lib/warning--filled/16';
-import PasswordInput from './PasswordInput';
-import ControlledPasswordInput from './ControlledPasswordInput';
+import PasswordInput, { ControlledPasswordInput } from './PasswordInput';
 import { textInputProps } from './util';
 
 const { prefix } = settings;
