+++ src/components/TextInput/index.js
@@ -6,6 +6,6 @@
  */
 
 export * from './TextInput.Skeleton';
-export ControlledPasswordInput from './ControlledPasswordInput';
+export ControlledPasswordInput from './PasswordInput';
 export PasswordInput from './PasswordInput';
 export default from './TextInput';
